pyproject.toml
[tool.poetry]
name = "fa-skeleton"
version = "0.1.0"
description = "A FastAPI project with AI/ML integrations for Expense Tracking"
authors = ["Your Name <your.email@example.com>"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "ai" },
    { include = "core" },
    { include = "middleware" },
    { include = "auth" },
]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.100.0"
uvicorn = "^0.23.0"
sqlalchemy = "^2.0.0"
alembic = "^1.12.0"
pydantic-settings = "^2.12.0"
aiosqlite = "^0.21.0"
greenlet = "^3.3.0"
pytest = "^9.0.2"
pytest-asyncio = "^1.3.0"
faker = "^38.2.0"
factory-boy = "^3.3.3"
pytest-cov = "^7.0.0"
mypy = "^1.19.0"
ruff = "^0.14.9"
black = "^25.12.0"
isort = "^7.0.0"
pre-commit = "^4.5.0"
safety = "^3.7.0"
detect-secrets = "^1.5.0"
passlib = "^1.7.4"
python-jose = "^3.5.0"
jupyter = "^1.1.1"
bcrypt = "<4"
asyncpg = "^0.31.0"
psycopg2 = "^2.9.11"

[tool.poetry.group.dev.dependencies]
httpx = "^0.24.1"

[tool.poetry.scripts]
# Remove the dev script - use 'poetry run uvicorn main:app --reload' instead

[tool.coverage.run]
omit = [
    "tests/*",
    "*/tests/*",
    "*/__pycache__/*",
    "*/site-packages/*"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
explicit_package_bases = true
mypy_path = "$MYPY_CONFIG_FILE_DIR"

[[tool.mypy.overrides]]
module = [
    "passlib.*",
    "jose.*",
    "factory.*",
    "faker.*",
]
ignore_missing_imports = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
